This patch introduces a configure switch --enable-virtualshapefunctions which will allow to replace the current Barton-Nackman class hierarchy by a more traditional one based on virtual functions.
  *** Unless this switch is set nothing changes ***

I admit that adding preprocessor code into this module is not a very
elegant thing to do.  However I believe there are good reasons to do so.

We need some form of dynamic polymorphism to be able to handle grids with
more than one element type.  Also, dynamic polymorphism creates simpler
compiler errors and makes code easier to read.  Of course the last point is
void if the preprocessor is used to switch between static and dynamic
polymorphism.

The main argument against dynamic polymorphism is the efficiency loss due
to virtual function calls.  While personally I believe that efficient assemblers
can be written even with a polymorphic class hierarchy, this question can
really only be answered with empirical data.  The new configure switch
will allow us to get such data.  In the short run it also allows people who
need dynamic polymorphism to have it without getting in the way of those
who insist on Barton-Nackman for top speed.


[[Imported from SVN: r70]]