Added template method evaluate to the LocalBasisVirtualInterface
Now the interface is compatible with static FEs.

In order to make double virtualization robust the virtual
evaluate method in the wrapper had to use either the template method
evaluate<k>(directions, ...) from a wrapped static LB or the virtual method
evaluate(directions, ) from a wrapped dynamic FE.

Unfortunately gcc<4.4 seems not to be able to determine the template
arument 'k' from the type of the argument 'directions' automatically.

Now that the interface has the template method it is always save
to specify the template argument in the wrapper.

[[Imported from SVN: r743]]