[Global FE] Rework the stuff about the orientation one more time.
Basically there is need for two types of orientation: tangential orientation
for dim=1 entities (edge elements) and normal orientation for codim=1 entities
(Raviart-Thomas).  Lines can be oriented tangentially by comparing the
numbering of their vertices in local indices of the reference element to the
numbering in the grid (global ids).  For normal orientation this not not
possible without further information -- best option seems to be comparing the
global ids of the neighboring elements to obtain the global orientation.

[[Imported from SVN: r896]]