[Derivatives] Convert from Int2Type to integral_constant.
All uses of Int2Type here appear to be internal to the class.  While there the
method tensor() which took Int2Type as dummy parameter _is_ public it does not
appear to be the intented external interface, since the class exists in two
implementations for different layouts ("value" vs. "derivative") and only the
"value" incarnation had the tensor(Int2Type) method anyway.

[[Imported from SVN: r917]]