[cmake,release] Install cmake macros to $(datadir)/dune/cmake/
Previously, the macros were installed to $(datadir)/cmake/. As we have
some macros in DUNE that might also be in other packages
this may cause conflicts in package managers. With this patch we install
them to $(datadir)/dune/cmake to resolve this issue. Together with the
last patch this will resolve problems  outlined in flyspray
task 813
https://dune-project.org/flyspray/index.php?do=details&task_id=813
I apologize for givin the wrong task number in the previous patch.
(cherry picked from commit ece1e584476965bb779f6ace315cc266d3b61306)

Signed-off-by: Markus Blatt <markus@dr-blatt.de>