[test] add test for diffOrder
in the documentation we state the diffOrder is set to >= 1, if evaluateJacobian can called.
All tests that test evaluateJacobian, now also require that diffOrder >= 1 is set. If the jacobian
test is disabled we expect the diffOrder to be 0, as the evaluateJacobian is either not
implemented or broken; at least it should not be called by the user.