Remove refinedp1test
It tested whether the refinement used for the 3d tetrahedral refinedP1 elements
matched the refinement used by ALUGrid.  While this was helpful in the beginning,
there is generally no reason why the RefinedP1Element should be consistent with
ALUGrid in particular.

Removing this test allows me to completely remove the dependency of dune-localfunctions
on dune-grid.