[bugfix] Properly implement the DisableJacobian check in all cases
Patch 5a34c1f5 introduced a few more tests depending on the DisableJacobian flag.
However, that patch was buggy, all it did was test whether the constant value
DisableJacobian ('3', currently), was true or not.  This patch fixes the issue.

Now most tests still fail, but that's apparently a separate issue: The static
diffOrder fields appears to be set to 0 for most finite elements, even the
differentiable ones.