Introduce method 'size' on the LocalFiniteElement class
It returns a (uint) number giving the number of shape functions/coefficients
of the finite element.  This number is of course also available through the
localBasis() and localCoefficients() methods, but having it directly on the
LocalFiniteElement class is very convenient.

Implementing this feature was decided at the 2013 Dune developer meeting
in Aachen.

This patch also extends test-localfe to test for the new method.