[bugfix] Remove check for diffOrder=1
This check was based on the false assumption that diffOrder
is related to the differentibility order. However it
actually means the order of implemented partial derivatives,
i.e. the max length of direction vectors you can pass to evaluate().

BTW this makes test-localfe pass again, while it failed for
almost all element types before.

In fact only the p1 element and the monomials implement
the evaluate method at all and correctly set the diffOrder.