[bugfix,test] Fix incorrect explicit enum value
The file contains an enum type whose values are used to disable individual
subtests.  The possible values are DisableNone, DisableLocalInterpolation,
DisableVirtualInterface, and DisableJacobian.  They were given the explicit
values 0,1,2,3 respectively.  However, in the code they are used to access
individual bits of the input argument 'char disableTests'.  Hence the values
need to be 0,1,2,4 instead.