[Bugfix][OSX] only call feenableexcept() on Linux
feenableexcept() is a Linux extension and doesn't exist on OS X, so
don't try to call it there.

Apart from that, we should think about just losing the call to
feenableexcept, as it only affects the x87 FPU and not SSE (and on 64
bit, modern GCC only generates SSE code anyway).