[cleanup] Get rid of 'evaluate()' in virtual interface
Since evaluate(...) for partial derivatives is gone, we can
clean up the virtual interface considerably. Now we have a single
interface layer instead of a hierarchy with one layer per diffOrder.

This also drops the LowerOrderLocalBasisTraits, because they where
only needed as implementation details of the implementation and the wrapper.

Since diffOrder is now completely useless (but still present in
LocalBasisTraits), we canonicalize its value to 0 using a template
alias. As a consequence LocalFiniteElementVirtualInterface<LB>
is now the same interface regardless of the diffOrder value in LB.

Without this trick you may run into trouble because you have different
interfaces depending on the value of diffOrder.