Hard-wire the maximal test order for 'partial' in the tests
Rather than using the value LocalBasisTraits::diffOrder.  That latter
value is scheduled for removal, and there is currently no replacement.
One of the reasons why there is no replacement is that nobody has a
real use-case for this information, other than unit testing.  To avoid
rushing out a new interface method only for testing, this patch hard-
wires the information in the test itself.