[cleanup] Reorder loops in TestEvaluate
Currently the evaluation of the differential quotient is done
for any shape function again. Since our interface does bulk-evaluation
of all shape functions at once, we effectively evaluate at the same
position n-times, where n is the number of shape functions.

This commit reorders the loops such that we only loop over
all shape functions in one of the inner loops and avoid this
problem. For a non-optimized build of testgerericfem this
reduces the runtime by ~25%.

Most of the diff is whitespace changes.