Rewrite testLocalInterpolation to match Ciarlet's definition
Previously, the testLocalInterpolation method constructed
random linear combinations of the shape functions, and then
checked whether the 'interpolate' method would recover the
random coefficients.  This patch rewrites the test to call
'interpolate' on the individual shape functions, and check
whether the resulting coefficient set is the correct unit
vector.  While this does not really result in an increase
of test coverage, it puts the test more in line with the
classic Ciarlet definition of a (local) finite element.
There, the shape functions are defined as the dual basis
to the set of degrees of freedom, and the new test tests
this directly. It also is a bit shorter, too.