Remove the diffOrder parameter completely
This patch completes the removal of the diffOrder template parameter.
The design of this was broken from its very beginning.  It played
very badly with the virtual interface, and it is not known whether
anybody really ever used it.  Its removal has lead to a vast
simplification of the dynamic interface.

Note that after this patch, there is no more way to find out
which partial derivatives are implemented by the 'partial' method
and which aren't.  Carsten and I decided to not implement and
interface for this before meeting the first person who actually
needs this.  Without such a person I for myself am not able to
design a good interface.