Add helper functions to translate between evaluate() and operator()
The provided functions allow to wrap functions supporting
one interface to model the other one. Using a concept check
this just forwards functions that already support the desired interface.

This is just an implementation detail for a transition process.
The purpose is to simply add a single line to each interpolate()
implementation in order to support both syntaxes.