Add support for operator() in interpolate
This is a first minimal version of supporting operator() in interpolate.
Working as follows:

* The virtual interpolate() method still uses VirtualFunction
* The templated interpolate() method does the following
  * Canonicalize the given function to operator() syntax
  * Wrap the result into a wrapper deriving from VirtualFunction
  * Pass this to the virtual interpolate()

This is close to the old implementation and has the benefit that
we don't need to touch the LocalInterpolationVirtualWrapper. I
will switch this follow up commits to a nicer version closer
to what we want after the transition:

* The virtual interpolate() method should be switched to std::function
* Any passed function should be wrapped to support operator() for
  transition