[bugfix] Fix function used in test
Some of our FE implementations require that the passed function
export a 'Traits' type providing the range and domain. Hence we should
also have this here.

Background: If you want to wrap a function supporting evaluate(x,y)
into operator() notation, you must create a 'y' and thus need to now
it's type. Some LFE implementations do fix this type manually but
extract it from the traits, such that we cannot now it if the
function does not support it.