Add support for operator() in interpolate() of meta/power
Which `operator()` one could replace `ComponentEvaluator` by
a simple lambda. But we cannot change this now without breaking
PowerInterpolation for FE-implementations outside of dune-localfunctions
which may not have been adjusted.

For the time being we have to derive `ComponentEvaluator` from
`Dune::Function` to ensure that it exports `Traits::DomainType`
and `Traits::RangeType`.