[Edge] Get a new reference element each time the element is constructed.
The previous approach used static class members to store the reference
elements at program startup.  This does not work, because the order of
initialization is unspecified if the class is a template class.  So instead,
fetch a new reference element each time the class is constructed, and hope
that that is cheap enough.

The alternate approach would have been to use a singleton returned by
reference from an `instance()` method.  That, however, would incur
synchronization.  The `instance()` method could have made the singleton
`thread_local` instead of `static`, but thread_local variables aren't cheap
either.

The current approach probably incurs synchronization, too, depending on what
`referenceElements()` does internally, but there is little that can be done
about it.