Make LocalFiniteElementVariant constructor less greedy
Avoid that the contructor can be instantiated with types
not contained in the implementation list. Otherwise this
may lead to nasty problems:

A std::vector<Std::optional<LocalFiniteElementVariant<...>>> cannot
be resized, because:

a)std::vector<...>::resize() checks for copy-constructability of its
  value type.
b)Our Std::optional has a constructor which is enabled_if the argument U
  is convertible to the wrapped type. In the check of a) U is itself
  an Std::optional
c)The check in b) tries to instantiate the constructor of the wrapped type
  from U.
d)As a consequence the constructor of LocalFiniteElementVariant<...>
  is instanciated with a Std::optional<LocalFiniteElementVariant<...>>
  as argument. Unsurprisingly this instanciation fails.

Remember: Always try to guard your greedy templated constructors properly.

With this patch the LocalFiniteElementVariant and LagrangeFiniteElementCache
seem to work nicely with std::optional and std::variant replaced by our own
Std::optional and Std::variant, respectively. Here Std::variant was taken
from a rebase-to-master version of:
https://gitlab.dune-project.org/core/dune-common/merge_requests/342