Make LocalFiniteElementVariant default-constructible
`LocalFiniteElementVariant` and its `LocalFooVariant` members
now inject the additional type `Std::monostate` into the
internally stored `Std::variant`s allowing to make
`LocalFiniteElementVariant` default constructible.
A default constructed
`LocalFiniteElementVariant` will always store a `Std::monostate`
and not an object of the first passed type, even if the latter
is default-constructible.

Notice that `Std::monostate` is automatically injected and
it is not supported to additionally pass it manually. I.e.
all passed types should be valid and compatible `LocalFiniteElement`
implementations.

As a side-effect the implementation is also cleaner and probably
more efficient:
We now use a single `Std::visit` that sets up all the members
of `LocalFiniteElementVariant` instead of separate `Std::visit`
calls for all members of all `LocalFooVariant` objects.