Workaround a bug in gcc 5 and gcc 6
Having multiple `using BaseClass::operator()` in a class
with multiple base classes does not work properly with gcc 5
and gcc 6 if `BaseClass` is a captureless lambda. This seems
to be related to the conversion-to-function-pointer operator
that is generated in this case.

While this can be avoided in gcc 6 by using a `[&]` although
nothing is captured at all, this workaround does not do the trick
in gcc 5. However, capturing a dummy value works for both
compilers.