New class LagrangeCubeLocalFiniteElement
This new class is intended to unify the existing implementations for
Lagrange elements on cube domains with compile-time dimension and
order.  This will make the API simpler (less implementations classes),
and the code gets shorter, too.

This commit also replaces the existing Q1LocalFiniteElement
implementation by the new one based on LagrangeCubeLocalFiniteElement.
This happens in a completely backward-compatible manner.