Partially bring back Q1LocalFiniteElement for older compilers
With the recent reimplementation of Q1LocalFiniteElement using
the new LagrangeCubeLocalFiniteElement, the unit test program 'test-q1.cc'
cannot be compiled with gcc-5 or gcc-6: The compiler simply starts
grabbing more and more memory, until the machine stalls.

This patch brings back a part of the old Q1LocalFiniteElement
implementation for these cases.