Deprecate duck-typed interface method from Lagrange element
The constructor of the Lagrange element which is provided a
vertex permutation previously accepted this in a duck-type
fashion through a template parameter. The reimplementation
by @oliver.sander internally works with std::array<unsigned int, dim + 1>
which seems the correct call to me as well.

I therefore propose to fix this type in the interface, deprecating
the old constructor.
Most likely, only maintainers of discretization packages are affected by this change.