Allow to skip test points when testing derivatives
There are finite element shape functions that are only piecewise
differentiable.  One example are the Lagrange shape functions
for pyramid elements, which are nondifferentiable on the plane
where x[0]==x[1].  Previously, the derivative tests where skipped
entirely for such elements.

This patch adds the infrastructure to skip individual test points
instead when testing derivatives.  This allows to test the
derivates where they are defined.  The points to skip are
specified by a predicate class that receives a point position
in the reference element, and returns 'true' if that point
is to be skipped.

This patch also uses the new infrastructure to enable the tests
for the derivatives of pyramidal Lagrange functions.