Add test whether shape functions can represent constants
This patch adds a new test that checks whether the functions space
spanned by the shape functions includes the constant functions.
This holds for all finite elements we currently implement.

This test can be disabled by setting DisableRepresentConstants.

The patch does disable the new test for the second order
Lagrange finite elements on the pyramid element, because
that one is apparently buggy.  A fix will come in a separate commit.