Fix buggy shape function for pyramids
One of the shape functions of the second-order Lagrange element for
pyramids was implemented wrongly, and the derivative along with it.
This went unnoticed for a long time, until I saw that the element
was not able to represent constant functions.

(cherry picked from commit 976bed3405ef6a1296f132d8524da49c43d4efe8)