Deprecate LocalFiniteElementFunctionBase
This class was needed to determine the base class to be used
for functions passing to `interpolate()`. Now we can simple
pass a callable object and type erasure in the virtual interface
is done using `std::function`.
Passing functions with the old interface providing an
`evaluate(x,y)` method is still possible using a wrapper.

To keep code using `LocalFiniteElementFunctionBase` working
during its deprecation, it exports a dummy base class which
is a look-a-like of the old `Dune::Function` base class.
This allows to no longer rely on the deprecated `Dune::Function`
class even during the deprecation period of
`LocalFiniteElementFunctionBase`.