generalize interface switch
The interface switch allows to handle local and global valued basis in
a generic fashion. Up to now the finite element was passed in as a
reference. We generalize the interface also allow rvalue references
and the storage object itself. This is a necessary change to support
on-the-fly generate basis objects in disretization modules like
PDELab.