[bugfix] Make static members constexpr
`static const` members need a definition at namespace
scope if ODR-used. Here the static member `dim` is
passed to a stream which may (or may not) be ODR-use.
The fact that it works so far is a lucky conincidence
since `DUNE_THROW` happens to be a non ODR-use.
Since C++17 `static constexpr` members are implicitly
inline and don't need a namespace scope definition
even if ODR-used.