[bugfix] Fix monomial basis on pyramids
Commit d6e8bbe from merge request
https://gitlab.dune-project.org/core/dune-localfunctions/-/merge_requests/190
broke the monomials (and thus the derived Lagrange basis)
by removing the special case for pyramids and thus the
Duffy transform. This patch reverts the parts of the
commit affecting the header `monomialbasis.hh` to fix
the problem. Additionally this incorporates the minor
changes applied to the header after the offending commit.