Implement arbitrary order partial derivatives in LagrangeSimplexLFE
We can implement arbitrary order derivatives by a simple recursion.
Notice that the effort grows exponentially with the total order because
the number of terms from the product rule is 2^totalOrder. Furthermore
the stack-storage grows linearly with the dynamic total order. Hence
we need to use a dynamic-to-static switch using an explicit enumeration
of the supported static derivative orders. Since derivatives of order >k
are zero, the number of these cases is known at compile time.