[bugfix] Replace the guessed RT-pyramid basis by the one by Ainsworth and Fu
The implemented basis is not a proper RT basis. It is probably not
in H(div) and cannot be glued together continuously with the tetrahedral
basis.

To fix this we now use the composite element proposed by Ainsworth/Fu in
"A lowest-order composite finite element exact sequence on pyramids"
Computer Methods in Applied Mechanics and Engineering,
Volume 324, 1 September 2017, Pages 110-127.
https://arxiv.org/abs/1705.00064

Notice that this is a composite element with a discontinuity across
the plane with x[0]==x[1]. It contains on DOF per face and one internal
DOF located at the center of the intersections of the two sub-elements.

In order to match the tetrahedral RT-basis, we use an uncommon scaling
of the triangluar facet normal momentum by an additional factor sqrt(2).
If the factor is removed in the tetrahedral RT basis, it should be
removed here, too.