[bugfix] Fix scaling of tetrahedral RT basis functions
The degrees of freedom are the normal facet moments.
This requires to scale the evaluation at the facet
center by the facet area as documented internally.
However, the values for the facet areas where incorrect
leading to the wrong scaling of basis functions.
Since the scaling was nevertheless consistent between
local basis and local interpolation is was not found
by the test.